src/main/java/com/developerdreamteam/jia/auth/rest/UserController.java
package com.developerdreamteam.jia.auth.rest;

import com.developerdreamteam.jia.auth.model.dto.UserDTO;
import com.developerdreamteam.jia.auth.model.entity.User;
import com.developerdreamteam.jia.auth.response.ServiceResponse;
import com.developerdreamteam.jia.auth.service.UserService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/api/v1/users")
public class UserController {

    @Autowired
    private UserService userService;

    @PostMapping("/signup")
    public ResponseEntity<ServiceResponse<User>> signup(@RequestBody UserDTO userDTO) {
        ServiceResponse<User> response = userService.saveUser(userDTO);
        return new ResponseEntity<>(response, response.getStatus());
    }
}
